works

+++ Server/src/main/java/via/sep4/model/Sensor/SensorRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 
-public interface SensorRepository extends JpaRepository<Sensor, Integer> {
+public interface SensorRepository extends JpaRepository<Sensor, Long> {
 }
